+++ src/control_plane/migration/object_store/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+mod tests;
+
+mod backup_minio_bucket;
+mod minio_backup_options;
+
+pub(crate) use backup_minio_bucket::backup_minio_bucket;
+pub(crate) use minio_backup_options::MinioBackupOptions;